fix Installer.edex-configuration %files

+++ rpms/unidata/build.sh
@@ -1,7 +1,9 @@
 #!/bin/bash -v
 set -xe
-rm -rf /awips2/jenkins/buildspace/workspace/AWIPS2-UPC_build/baseline/
-export JENKINS_WORKSPACE=/awips2/repo/awips2-builds
+if  [ ${2} = "buildCAVE"]; then
+  rm -rf /awips2/jenkins/buildspace/workspace/AWIPS2-UPC_build/baseline/
+fi
+export JENKINS_WORKSPACE=/home/awips/awips2-builds
 
 # determine where we are ...
 path_to_script=`readlink -f $0`
